fix: Move layout file to its own file
